1.Double-Click bugfix: access item by proxy  2.Renamed classes
