+++ plugins/gui/include/gui/context_manager_widget/models/context_tree_model.h
